CatsAndDogs: Setup loss calculations
-- This is a rebase commit message, the rest here are messages
from relevant previous commits in ascending order of date --

- Add reshape utilities

Decided to put it in a separate file
as it's all quite messy, and I thought
there might be a number of different
reshape functions, but now I'm actually
keeping it quite simple, for now anyway.

Might want to ultimately move a bunch
of stuff to separate files as the main
notebook's quite messy.

- Add function for calculating ground truth classes

Working towards actually being able to calculate the losses.
Keeping it fairly simple again, only considering batch sizes
of one. Would rather get it to a working state first.

- Setup function for easy loss calculation