CatsAndDogs: Add basic train loop, testing for predictions
Just to get it working somewhat for now. Lots of issues
still, but testing with a small dataset, it does seem
to at the very least be learning, seeing as the
loss does drop. Don't really know if both losses
are dropping though, might want to actually do them
as separate because of that.

Two things to do are hard negative mining for culling
the number of negative default box examples, and
non-maximum suppression after the predictions are made.
It seems both need confidences for the boxes, something
that isn't seemingly by default part of the predictions
based on the description of the SSD, so not sure what's up
with that. Should of course
not be too difficult to add confidences, as it's just
another channel per box.