CatsAndDogs: Fix box offset calculation
Was scaling the boxes wrongly, should be
good now.

Also changed to just using one example in training for now,
as at the very least the network should be able
to basically match the one example exactly.
Fairly poor prediction performance on just the one
example though, so clearly need to improve
quite a bit.