CatsAndDogs: Add "hard negative mining"
This should, and kind of seems to,
improve the class predictions, at least in
the sense that it's now a bit better
at giving other predictions other than just
the background, while still seemingly
giving decent background predictions as well.

I did it based on what the
true classes are, which does create the
difficulty that the false predictions made
of other classes may be underrepresented in
training. I'll also try picking the background
examples from the predicted examples instead,
which should allow keeping other, non-background
predictions in training a little better.