CatsAndDogs: Change to using a combination of two different class losses
As mentioned in the notebook, the idea
is that using hard negative mining
based on just the true or predicted
classes is perhaps not as useful
for learning as just using both
simultaneously. It certainly seems
to work okay this way, where I do
feel like it's able to still
slightly better predict the positive
classes, while not having a massive
amount of them, something that
seemed to perhaps be an issue with
using just the true classes.

Might also want to add the ability to
weight the two losses.