CatsAndDogs: Change to using CUDA distribution of torch
Decided it would be nice to try with CUDA as well,
as the CPU training is fairly slow.

Need to still actually include the use of the GPU
in the code, which will probably take a bunch of
fiddling. Obviously would have again been good
to include the use of the device from the get-go.