CatsAndDogs: Add option for using CUDA in model
GPU seems to be about three times as fast as CPU,
currently, which is decent but not amazing. Should
be able to improve the speed if I can find stuff
to vectorise more, which I imagine I could.

The main thing currently standing in the way of
training more examples is the IoU calculation
as it is very slow. I would imagine I would need
to improve it quite a bit though so that it can
be run each iteration with not-too-big overhead,
as I'm not entirely sure each IoU can actually
be cached (memory-wise), though could obviously
try (or calculate).
Still, I do want to improve the speed regardless.

Second thing is better batch processing of
the training examples, namely the ability
to process all of a batch in one forward
pass in a vectorised way.
Not sure how implementable this
is in the current state though, have to
think about it a little.