CatsAndDogs: Make predictions work in batched form
Wanted to start moving the calculation to a more
batched form to see if the training could
be sped up this way. The only thing
that's properly vectorised is the predictions,
though, with losses just calculated in a loop.
It doesn't seem like there's any considerable
change in training speed as a result, though
I did also change to calculating the IoUs each
time the loss is calculated. Either way,
I'm hoping for more considerable improvements
to the training speed.

It might certainly be possible to batch
some other calculations too. The first
thing to look at is the IoU calculation,
I think.