CatsAndDogs: Make a batched version of IoU
At least more batched than previously,
and it does work faster by some amount.
It currently works by having the
batches in lists, as it would otherwise
be somewhat confusing: could have it as
a flat tensor, but it's more difficult
to keep track of which indices make up
which batch. This does mean that there's
currently overhead in having to constantly
move the tensors back and forth from
the list form between calls, at least
once other functions are also brought
to a batched form. The assumption is,
however, that this overhead is smaller
than the overall gain in speed.

Might of course be sensible to do
some profiling of the code to not
concentrate on inconsequential issues,
though at the same time, it's good
practice to do some batching of the
functions too.