CatsAndDogs: Remove unnecessary .to transformations
The .to transforms are a major contributor
to how long the training takes, so I decided
to at least get rid of some of the ones
that were still hanging around without being
particularly useful.

This did not make much of a difference, so
the real culprits are somewhere else, potentially
the ones that change to and from GPU in the forward
function. Another major contributor to the time
is the loss calculations. Both this and the
transformations seem to independently take roughly
the same time as the backward pass calculation.
In order to get rid of some of the transforms,
the calculations would have to be done more on
GPU in general. However, for this to be ideal,
the loops in the loss calculations would have
to be vectorised, as currently running the whole
thing on the GPU is very slow compared to
just the model on GPU and the rest on CPU.