CatsAndDogs: Use test loss in training
Decided it'd be nice to finally add
the proper setup for the use of
test and validation data, too, though
need to actually do proper testing
on the validation data still.
Other things are properly batched
training data, and potentially
non-maximum suppression, but that's
not really that relevant to the
problem.

Also, could be useful
to separate the training of the
two networks (box and class
predictions) as they are effectively
completely independent already, and
their training does not happen
at the same rate.