CatsAndDogs: Add proper evaluation for classes
The boxes are more difficult to evaluate
in a sensible way, besides just using
the loss but that doesn't really
give an intuitive idea of the performance.
Ideally you'd have non-maximum suppression
to properly find a set of "good" guesses,
then compare the predicted and true class
and IoU against the matched box to determine
the accuracy.