imdb: Fix duplicate column issue
Using "context" as the column for the positive/negative
naturally overlapped with the actual word "context"
that was in the training data, which caused problems.
Although, there was some weirdness with it being seemingly
suggested that there were no non-unique column names, yet
parquet complained? Still, seems to be working now.