imdb: Move dataset class to separate file
Largely the same content in the classes that were in the
two files previously, so move to one place to make it easier
to add stuff to the class. Could also do an abstract class
to make interfacing more clear, but it's not really that
important with this ad-hoc stuff right now.

At the same time, fixed the context mapping in the class.
This wasn't being used yet, so didn't really matter, but
also added true/false negative/positive rate calculations,
so it was necessary for that.